// src/tasks/enum/task-type.enum.ts
// src/tasks/task-type.enum.ts
export enum TaskType {
  VISIT_WEBSITE = 'VISIT_WEBSITE',
  FOLLOW_SOCIAL = 'FOLLOW_SOCIAL',
  WATCH_YOUTUBE = 'WATCH_YOUTUBE',
  JOIN_TELEGRAM_CHANNEL = 'JOIN_TELEGRAM_CHANNEL',
  BOOST_TELEGRAM_CHANNEL = 'BOOST_TELEGRAM_CHANNEL',
  POST_TELEGRAM_STORY = 'POST_TELEGRAM_STORY',
  ADD_LOGO_TO_PROFILE_NAME = 'ADD_LOGO_TO_PROFILE_NAME',
  CONNECT_WALLET = 'CONNECT_WALLET',
  ON_CHAIN_TRANSACTION = 'ON_CHAIN_TRANSACTION',
  ADD_TOKEN_TO_WALLET = 'ADD_TOKEN_TO_WALLET',
}